Progrès: Ajout de vote dans la base de donnée
Ajout de vote dans la base de donné fonctionnel, mais relié a aucune question. Ajout des questions ajouté aussi pusique j'ai oublié de stage les files.

+++ app/src/main/java/felix/gagne/votodroidgagne/modele/Vote.java
@@ -12,7 +12,7 @@ public class Vote {
     public Long id;
 
     @ColumnInfo
-    public Long value;
+    public float value;
 
     @ColumnInfo
     public String nomDuVotant;
